Move dashboard files to dedicated folder
