+++ app/(dashboard)/issues/[id]/IssueDetails.tsx
