+++ app/(dashboard)/issues/[id]/page.tsx
