+++ app/(dashboard)/issues/edit/[id]/page.tsx
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic'
 import IssueFormSkeleton from './loading'
 
 const IssueForm = dynamic(
-  () => import('@/app/issues/_components/IssueForm'),
+  () => import('@/app/(dashboard)/issues/_components/IssueForm'),
   {
     ssr: false,
     loading: () => <IssueFormSkeleton />
