+++ app/(dashboard)/issues/list/loading.tsx
