+++ app/(dashboard)/issues/new/page.tsx
@@ -2,7 +2,7 @@ import dynamic from "next/dynamic"
 import IssueFormSkeleton from "./loading"
 
 const IssueForm = dynamic(
-  () => import('@/app/issues/_components/IssueForm'), 
+  () => import('@/app/(dashboard)/issues/_components/IssueForm'), 
   {
     ssr: false,
     loading: () => <IssueFormSkeleton />
