+++ package.json
@@ -20,11 +20,13 @@
     "@types/react-dom": "18.2.15",
     "autoprefixer": "10.4.16",
     "axios": "^1.5.0",
+    "bcryptjs": "^2.4.3",
     "classnames": "^2.3.2",
     "delay": "^6.0.0",
     "easymde": "^2.18.0",
     "eslint": "8.54.0",
     "eslint-config-next": "14.0.3",
+    "jsonwebtoken": "^9.0.2",
     "moment": "^2.30.1",
     "next": "14.0.3",
     "next-auth": "^4.24.5",
@@ -44,6 +46,8 @@
     "zod": "^3.22.2"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.10"
+    "@tailwindcss/typography": "^0.5.10",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/jsonwebtoken": "^9.0.5"
   }
 }
